And fix the other one too

+++ code/object/object.cpp
@@ -519,7 +519,7 @@ int obj_create(ubyte type,int parent_obj,int instance, matrix * orient,
 void obj_delete_all() 
 {
 	int counter = 0;
-	for (size_t i = 0; i < MAX_OBJECTS; ++i) 
+	for (int i = 0; i < MAX_OBJECTS; ++i) 
 	{
 		if (Objects[i].type == OBJ_NONE)
 			continue;
